refactor: move BitcoinTx into bitcoin
